refactor: 모듈 infra -> auth-infra로 변경
