+++ auth/auth-infra/src/main/java/shop/woowasap/auth/infra/configuration/WebSecurityConfiguration.java
@@ -1,4 +1,4 @@
-package shop.woowasap.infra.configuration;
+package shop.woowasap.auth.infra.configuration;
 
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
@@ -18,7 +18,7 @@ import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 public class WebSecurityConfiguration {
 
     @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+    public SecurityFilterChain securityFilterChain(final HttpSecurity http) throws Exception {
         http
             .csrf(AbstractHttpConfigurer::disable)
             .httpBasic(AbstractHttpConfigurer::disable)
