+++ auth/auth-infra/src/main/java/shop/woowasap/auth/infra/filter/JwtTokenAuthFilter.java
@@ -1,4 +1,4 @@
-package shop.woowasap.infra.filter;
+package shop.woowasap.auth.infra.filter;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -7,6 +7,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.filter.OncePerRequestFilter;
+import shop.woowasap.auth.infra.JwtTokenProvider;
 
 @RequiredArgsConstructor
 public class JwtTokenAuthFilter extends OncePerRequestFilter {
@@ -16,7 +17,7 @@ public class JwtTokenAuthFilter extends OncePerRequestFilter {
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
         FilterChain filterChain) throws ServletException, IOException {
-        
+
         filterChain.doFilter(request, response);
     }
 }
