+++ coverage-exclude.asap
@@ -10,4 +10,5 @@ exclude shop/woowasap/*/app/exception/*
 exclude shop/woowasap/*/domain/*/request/*
 exclude shop/woowasap/*/domain/*/response/*
 exclude shop/woowasap/auth/service/configuration/*
-exclude shop/woowasap/infra/*/*
+exclude shop/woowasap/auth/infra/*/*
+exclude shop/woowasap/auth/infra/*
